requirements.txt
python-telegram-bot==21.7
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
